created a route to get all projects the routes checks  if there is no available projectand return a status of 404
