+++ app.js
@@ -1,4 +1,3 @@
-
 import express from 'express';
 import dotenv from 'dotenv';
 import cookieParser from 'cookie-parser';
@@ -6,6 +5,7 @@ import cors  from "cors"
 
 import authRouter from './routes/auth.js';
 import usersRouter from './routes/users.js';
+import projectRouter from './routes/projects.js';
 
 dotenv.config();
 const app = express();
@@ -21,6 +21,7 @@ app.use(cookieParser());
 
 app.use('/auth', authRouter);
 app.use('/users', usersRouter);
+app.use('/projects', projectRouter);
 
 // catch 404 and forward to error handler
 // app.use(function (req, res, next) {
