+++ utils/projectInputValidator.js
@@ -1,16 +1,18 @@
 import Joi from "joi";
 
-const projectSchema = {
+const projectSchema = Joi.object({
   name: Joi.string().min(3).max(255).required(),
   description: Joi.string().required(),
-  deadline: Joi.date().required()
-}
+  deadline: Joi.date().required(),
+});
+
 export default function projectValidator(req, res, next) {
   const { error, value } = projectSchema.validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message)
-    req.validatedProject = value
-  next()
+  if (error) return res.status(400).send(error.details[0].message);
+  req.validatedProject = value;
+  next();
 }
+
 // CREATE TABLE projects (
 //   id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
 //   name VARCHAR(255) NOT NULL,
